Move configuration parameter-specific tests to separate test class
